New Commits to Add Movie

+++ Models/Movie.cs
@@ -17,7 +17,7 @@ namespace Vidly.Models
 
         [Display(Name ="Genre")]
         [Required]
-        public byte GenreId { get; set; }
+        public int GenreId { get; set; }
 
         public DateTime DateAdded { get; set; }
 
@@ -26,6 +26,6 @@ namespace Vidly.Models
 
         [Display(Name = "Number in Stock")]
         [Range(1,20)]
-        public byte NumberInStock { get; set; }
+        public int NumberInStock { get; set; }
     }
 }
